+++ Sora/MediaChannel.swift
@@ -9,9 +9,6 @@ public final class MediaChannelHandlers {
     /// 接続解除時に呼ばれるブロック
     public var onDisconnectHandler: ((Error?) -> Void)?
     
-    /// 接続中にエラーが発生したときに呼ばれるブロック
-    public var onFailureHandler: ((Error) -> Void)?
-    
     /// シグナリングメッセージの受信時に呼ばれるブロック
     public var onMessageHandler: ((SignalingMessage) -> Void)?
     
@@ -205,14 +202,22 @@ public final class MediaChannel {
         state = .connecting
         connectionStartTime = nil
 
+        peerChannel.internalHandlers.onDisconnectHandler = { error in
+            if self.state == .connecting || self.state == .connected {
+                self.disconnect(error: error)
+            }
+        }
+        
         peerChannel.internalHandlers.onAddStreamHandler = { stream in
             Logger.debug(type: .mediaChannel, message: "added a stream")
+            Logger.debug(type: .mediaChannel, message: "call onAddStreamHandler")
             self.internalHandlers.onAddStreamHandler?(stream)
             self.handlers.onAddStreamHandler?(stream)
         }
         
         peerChannel.internalHandlers.onRemoveStreamHandler = { stream in
             Logger.debug(type: .mediaChannel, message: "removed a stream")
+            Logger.debug(type: .mediaChannel, message: "call onRemoveStreamHandler")
             self.internalHandlers.onRemoveStreamHandler?(stream)
             self.handlers.onRemoveStreamHandler?(stream)
         }
@@ -221,6 +226,8 @@ public final class MediaChannel {
             Logger.debug(type: .mediaChannel, message: "receive event notification")
             self.publisherCount = message.publisherCount
             self.subscriberCount = message.subscriberCount
+            
+            Logger.debug(type: .mediaChannel, message: "call onNotificationEventHandler")
             let event = NotificationEvent(message: message)
             self.internalHandlers.onNotificationEventHandler?(event)
             self.handlers.onNotificationEventHandler?(event)
@@ -240,6 +247,8 @@ public final class MediaChannel {
                 Logger.error(type: .mediaChannel, message: "failed to connect")
                 self.disconnect(error: error)
                 handler(error)
+                
+                Logger.debug(type: .mediaChannel, message: "call onConnectHandler")
                 self.internalHandlers.onConnectHandler?(error)
                 self.handlers.onConnectHandler?(error)
                 return
@@ -247,6 +256,7 @@ public final class MediaChannel {
             Logger.debug(type: .mediaChannel, message: "did connect")
             self.state = .connected
             handler(nil)
+            Logger.debug(type: .mediaChannel, message: "call onConnectHandler")
             self.internalHandlers.onConnectHandler?(nil)
             self.handlers.onConnectHandler?(nil)
         }
@@ -270,17 +280,20 @@ public final class MediaChannel {
             
         default:
             Logger.debug(type: .mediaChannel, message: "try disconnecting")
+            if let error = error {
+                Logger.error(type: .mediaChannel,
+                             message: "error: \(error.localizedDescription)")
+            }
+            
             state = .disconnecting
             connectionTimer.stop()
             peerChannel.disconnect(error: error)
             Logger.debug(type: .mediaChannel, message: "did disconnect")
             state = .disconnected
+            
+            Logger.debug(type: .mediaChannel, message: "call onDisconnectHandler")
             internalHandlers.onDisconnectHandler?(error)
             handlers.onDisconnectHandler?(error)
-            if let error = error {
-                internalHandlers.onFailureHandler?(error)
-                handlers.onFailureHandler?(error)
-            }
         }
     }